refactor(macos): split terminal surface controller owners
