+++ clients/apple/scripts/check-architecture-maintainability.sh
@@ -204,7 +204,6 @@ current_root_swift_allowlist=(
     "ShellHostController.swift"
     "TerminalPaneView.swift"
     "TerminalRuntimeRegistry.swift"
-    "TerminalSurfaceController.swift"
 )
 
 target_dirs=(
@@ -608,7 +607,7 @@ while IFS= read -r file; do
 
     if grep -Eq '^import (AppKit|Darwin)$' "$file"; then
         case "$rel" in
-            App/*|Services/*|Support/*|Views/Shell/Terminal/*|AlanApp.swift|AlanAppSingletonGuard.swift|GhosttyLiveHost.swift|ShellControlPlane.swift|TerminalSurfaceController.swift)
+            App/*|Services/*|Support/*|Views/Shell/Terminal/*|AlanApp.swift|AlanAppSingletonGuard.swift|GhosttyLiveHost.swift|ShellControlPlane.swift)
                 ;;
             MacShellRootView.swift|ShellHostController.swift|TerminalRuntimeRegistry.swift)
                 warn "bridge|$rel|appkit-or-darwin-outside-bridge" \
